Test behavior of objects entering and exiting scope.
The cube is outside your visible distance to start with, so there should
only be 1 ghost in the net graph. If you move forwards, you should see
the cube, and the ghost count will increase to 2. Move backwards again,
so the cube leaves your visible distance, and the ghost count will stay
at 2.